Adds timer to question answering
Implements a timer for each question in answering mode.
The timer automatically submits the answer when it expires.
Also introduces test repositories and view models for testing.
Allows null selected choice ID to mark question viewed event.
Renames QuestionsView to QuestionsViewModel and moves it to ViewModels folder.
Deletes the ArcTimerButton swift file and moves it to the ViewComponents folder